+++ src/cmd/internal/robustio/robustio.go
