+++ src/cmd/internal/robustio/robustio_darwin.go
