+++ src/cmd/internal/robustio/robustio_other.go
