+++ src/cmd/internal/robustio/robustio_windows.go
